+++ plot_training_data.py
@@ -19,13 +19,13 @@ def plot_training_data(filenames, sampling_rate=48000):
   """
 
   for fn in filenames:
-    vecs = np.loadtxt(fn, delimiter=',', dtype=np.int16)
+    vecs = np.loadtxt(fn, delimiter=',')
     for i, vec in enumerate(vecs):
       plt.specgram(vec, Fs=sampling_rate)
       plt.title(i+1)
       button_axes = plt.axes([0.81, 0.05, 0.1, 0.075])
       button = Button(button_axes, 'Play')
-      button.on_clicked(lambda _: sd.play(vec, sampling_rate))
+      button.on_clicked(lambda _: sd.play(vec.astype(np.int16), sampling_rate))
       plt.show()
